pyproject.toml
[project]
name = "blendtutor"
version = "0.1.0"
description = "AI-powered feedback on interactive coding exercises via Fireworks AI"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
educator = ["deepeval"]

[project.entry-points."blendtutor.lessons"]
blendtutor = "blendtutor"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/blendtutor"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
src = ["src"]
line-length = 88

[tool.uv.workspace]
members = [
    "~/Documents/coding/portfolio/justenougheng-py",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.1",
]
